Minor fixes
Dont need to copy all files from overlay. Copying a an unmodified file
from overlay fs to lower results in 0 byte file. Don't copy
ovs-port-br-ex because of this. It wont be modified during update.

TODO see if we need to actually reboot, or can we just modify nbctl for
GR

Signed-off-by: Tim Rozet <trozet@redhat.com>