More fixes
We need to know the old interface because configure-ovs  will take down
the default gateway interface, which will be our old interface after
migration. We have to bring it back up with this script.

Also, by using a systemd service we dont need to reboot anything if we
can migrate OVS before kubelet and ovnkube-node starts. That's the
theory anyway...

Signed-off-by: Tim Rozet <trozet@redhat.com>